tests/system: split instrumentation tests out
